requirements.txt
pytest==7.4.3
pytest-playwright==0.4.3
playwright==1.39.0
allure-pytest==2.13.2
allure-python-commons==2.13.2
requests==2.31.0
python-dotenv==1.0.0
Pillow==10.1.0
numpy==1.26.2
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-html==4.1.1